# services/rest-api-svc/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Go Micro v5, NATS, PostgreSQL 和 Redis 构建的生产级短链接微服务。\n此 REST API 为终端用户提供 HTTP 端点，用于与短链接服务进行交互。",
        "title": "URL 短链接 REST API",
        "contact": {
            "name": "URL Shortener API 支持",
            "email": "support@urlshortener.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "192.168.0.51:8080",
    "basePath": "/api/v1",
    "paths": {
        "/analytics/top-urls": {
            "get": {
                "description": "根据点击量检索表现最出色的顶级短链接列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分析统计"
                ],
                "summary": "获取热门 URL 排行榜",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int32",
                        "example": 10,
                        "description": "返回的 URL 数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 1672531200,
                        "description": "开始时间 (Unix 时间戳)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 1672617600,
                        "description": "结束时间 (Unix 时间戳)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "clicks",
                        "description": "排序维度 (clicks/unique_clicks/created_at)",
                        "name": "sort_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top URLs retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.TopURLsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/analytics/urls/{shortCode}": {
            "get": {
                "description": "检索特定短链接的详尽分析数据（包括点击量、地理位置、设备分布等）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分析统计"
                ],
                "summary": "获取 URL 分析统计数据",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "短码标识符",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 1672531200,
                        "description": "开始时间 (Unix 时间戳)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "example": 1672617600,
                        "description": "结束时间 (Unix 时间戳)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "day",
                        "description": "时间粒度 (hour/day/week/month)",
                        "name": "granularity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL analytics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.URLStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/shorten": {
            "post": {
                "description": "根据长链接生成短链接，支持自定义别名和过期时间设置",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接管理"
                ],
                "summary": "创建短链接",
                "parameters": [
                    {
                        "description": "短链接生成请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ShortenURLRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created short URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ShortenURLResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/urls/{shortCode}": {
            "get": {
                "description": "根据短码检索短链接的详细元数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL 管理"
                ],
                "summary": "获取 URL 详情",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "短码标识符",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "user123",
                        "description": "用户 ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.URLInfoResponse"
                        }
                    },
                    "400": {
                        "description": "Missing user_id parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "URL not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除属于已验证用户的特定短链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL 管理"
                ],
                "summary": "删除短链接",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "短码标识符",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "user123",
                        "description": "用户 ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "URL deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Missing user_id parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete URL",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userID}/urls": {
            "get": {
                "description": "检索属于特定用户的短链接分页列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "获取用户的 URL 列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "user123",
                        "description": "用户唯一标识 ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 20,
                        "description": "每页显示条数",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "created_at",
                        "description": "排序字段（如 created_at, click_count）",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "desc",
                        "description": "排序顺序（asc 升序, desc 降序）",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User URLs retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/handler.UserURLsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get user URLs",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{shortCode}": {
            "get": {
                "description": "解析短码，跳转至原始长链接并记录点击追踪数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "重定向"
                ],
                "summary": "重定向至原始 URL",
                "parameters": [
                    {
                        "type": "string",
                        "example": "abc123",
                        "description": "短码标识符",
                        "name": "shortCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect information (for API testing)",
                        "schema": {
                            "$ref": "#/definitions/handler.RedirectResponse"
                        }
                    },
                    "302": {
                        "description": "Redirect to original URL"
                    },
                    "404": {
                        "description": "Short code not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "410": {
                        "description": "URL has expired",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.BrowserStatsItem": {
            "type": "object",
            "properties": {
                "browser": {
                    "description": "浏览器名称及版本号",
                    "type": "string",
                    "example": "Chrome 120"
                },
                "clicks": {
                    "description": "该浏览器的累计点击数",
                    "type": "integer",
                    "example": 80
                },
                "percentage": {
                    "description": "该浏览器占比百分比 (0-100)",
                    "type": "number",
                    "example": 53.3
                }
            }
        },
        "handler.CountryStatsItem": {
            "type": "object",
            "properties": {
                "clicks": {
                    "description": "点击量：来自该国家或地区的总访问次数",
                    "type": "integer",
                    "example": 75
                },
                "country": {
                    "description": "国家代码：通常使用 ISO 标准缩写（如 \"CN\" 代表中国, \"US\" 代表美国）",
                    "type": "string",
                    "example": "US"
                },
                "percentage": {
                    "description": "百分比：该地区的点击量占总点击量的百分比（0.0-100.0）",
                    "type": "number",
                    "example": 50
                }
            }
        },
        "handler.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "提示消息：说明操作的结果，例如 \"URL 已成功删除\" 或具体的失败原因",
                    "type": "string",
                    "example": "URL deleted successfully"
                }
            }
        },
        "handler.DeviceStatsItem": {
            "type": "object",
            "properties": {
                "clicks": {
                    "description": "来自该类型设备的点击总数",
                    "type": "integer",
                    "example": 90
                },
                "device_type": {
                    "description": "设备类型（如 Desktop, Mobile, Tablet）",
                    "type": "string",
                    "example": "Desktop"
                },
                "percentage": {
                    "description": "该设备点击量在总点击中的占比百分比",
                    "type": "number",
                    "example": 60
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "handler.RedirectResponse": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer",
                    "example": 43
                },
                "long_url": {
                    "type": "string",
                    "example": "https://www.google.com"
                },
                "session_id": {
                    "type": "string",
                    "example": "sess_1234567890abcdef"
                },
                "short_code": {
                    "type": "string",
                    "example": "abc123"
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1672531200
                }
            }
        },
        "handler.ReferrerStatsItem": {
            "type": "object",
            "properties": {
                "clicks": {
                    "description": "该来源贡献的累计点击数",
                    "type": "integer",
                    "example": 45
                },
                "percentage": {
                    "description": "该来源占总流量的百分比 (0-100)",
                    "type": "number",
                    "example": 30
                },
                "referrer": {
                    "description": "来源地址（如搜索引擎、社交媒体或特定网站）",
                    "type": "string",
                    "example": "https://google.com"
                }
            }
        },
        "handler.ShortenURLRequest": {
            "type": "object",
            "required": [
                "long_url",
                "user_id"
            ],
            "properties": {
                "custom_alias": {
                    "type": "string",
                    "example": "google"
                },
                "expiration_time": {
                    "type": "integer",
                    "example": 1735689600
                },
                "long_url": {
                    "type": "string",
                    "example": "https://www.google.com"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "campaign": "social",
                        "source": "twitter"
                    }
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "handler.ShortenURLResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "integer",
                    "example": 1672531200
                },
                "expires_at": {
                    "type": "integer",
                    "example": 1735689600
                },
                "long_url": {
                    "type": "string",
                    "example": "https://www.google.com"
                },
                "short_code": {
                    "type": "string",
                    "example": "abc123"
                },
                "short_url": {
                    "type": "string",
                    "example": "https://short.ly/abc123"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "handler.TimeSeriesPoint": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer",
                    "example": 25
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1672531200
                },
                "unique_clicks": {
                    "type": "integer",
                    "example": 18
                }
            }
        },
        "handler.TopURLItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "链接创建的时间戳",
                    "type": "integer",
                    "example": 1672531200
                },
                "last_clicked": {
                    "description": "最后一次被点击的时间戳",
                    "type": "integer",
                    "example": 1672617600
                },
                "short_code": {
                    "description": "短码标识符",
                    "type": "string",
                    "example": "abc123"
                },
                "total_clicks": {
                    "description": "总点击次数",
                    "type": "integer",
                    "example": 150
                },
                "unique_clicks": {
                    "description": "独立访客点击数",
                    "type": "integer",
                    "example": 95
                }
            }
        },
        "handler.TopURLsResponse": {
            "type": "object",
            "properties": {
                "urls": {
                    "description": "热门短链接信息列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TopURLItem"
                    }
                }
            }
        },
        "handler.URLInfoResponse": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer",
                    "example": 42
                },
                "created_at": {
                    "type": "integer",
                    "example": 1672531200
                },
                "expires_at": {
                    "type": "integer",
                    "example": 1735689600
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "long_url": {
                    "type": "string",
                    "example": "https://www.google.com"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "campaign": "social",
                        "source": "twitter"
                    }
                },
                "short_code": {
                    "type": "string",
                    "example": "abc123"
                },
                "short_url": {
                    "type": "string",
                    "example": "https://short.ly/abc123"
                },
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "handler.URLStatsResponse": {
            "type": "object",
            "properties": {
                "browser_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.BrowserStatsItem"
                    }
                },
                "country_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CountryStatsItem"
                    }
                },
                "device_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.DeviceStatsItem"
                    }
                },
                "referrer_stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.ReferrerStatsItem"
                    }
                },
                "short_code": {
                    "type": "string",
                    "example": "abc123"
                },
                "time_series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.TimeSeriesPoint"
                    }
                },
                "total_clicks": {
                    "type": "integer",
                    "example": 150
                },
                "unique_clicks": {
                    "type": "integer",
                    "example": 95
                }
            }
        },
        "handler.UserURLsResponse": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean",
                    "example": true
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total_count": {
                    "type": "integer",
                    "example": 100
                },
                "urls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.URLInfoResponse"
                    }
                }
            }
        }
    }
}